Countdown to devastation v1.0
In the late 1980's agents of the Ministry of Occultism are dispatched to the remote town of Axbury, where the Cult of Azathoth are plotting to use a bomb to open a weak point in our reality, and destroy the walls between our world, and that of the Blind, Idiot God. \n  \n This scenario is inspired by Differently Morphous and Existentially Challenged, by Yhatzee Croshaw.